+++ extensions/plg_user_pungamail/services/provider.php
@@ -11,6 +11,7 @@ use Joomla\CMS\Factory;
 use Joomla\CMS\Plugin\PluginHelper;
 use Joomla\DI\Container;
 use Joomla\DI\ServiceProviderInterface;
+use Joomla\Event\DispatcherInterface;
 use Punga\Plugin\User\PungaMail\Extension\PungaMail;
 
 return new class implements ServiceProviderInterface
@@ -22,7 +23,10 @@ return new class implements ServiceProviderInterface
 			PluginInterface::class,
 			static function (Container $container): PluginInterface
 			{
-				$plugin = new PungaMail((array) PluginHelper::getPlugin('user', 'pungamail'));
+				$plugin = new PungaMail(
+					$container->get(DispatcherInterface::class),
+					(array) PluginHelper::getPlugin('user', 'pungamail')
+				);
 				$plugin->setApplication(Factory::getApplication());
 
 				return $plugin;